// src/config/env.d.ts
export declare const config: {
    notion: {
        apiKey: string;
        databaseId: string;
    };
    discord: {
        token: string;
        channelId: string;
    };
    qbittorrent: {
        url: string;
        username: string;
        password: string;
    };
};
//# sourceMappingURL=env.d.ts.map